flowsdk::mqtt_client::engine

Type Alias MqttMessage

pub type MqttMessage = MqttPublish;

Alias for MqttPublish (v5) to provide a single, unified type for received messages.

The engine normalizes all incoming PUBLISH packets (whether MQTT v3.1.1 or v5.0) into this structure. This simplifies downstream consumption by providing a consistent API regardless of the protocol version used. For MQTT v3.1.1 messages, the v5-specific fields (Properties) will be empty.

Aliased Type

pub struct MqttMessage {
    pub topic_name: String,
    pub qos: u8,
    pub dup: bool,
    pub retain: bool,
    pub packet_id: Option<u16>,
    pub payload: Vec<u8>,
    pub properties: Vec<Property>,
}

Fields
topic_name: Stringqos: u8dup: boolretain: boolpacket_id: Option<u16>payload: Vec<u8>properties: Vec<Property>